Restructure protobuf files
