--- rpc/pb/generate.go
@@ -1,4 +1,0 @@
-package va
-
-//go:generate protoc --go_out=plugins=grpc:. core/core.proto
-//go:generate protoc --go_out=plugins=grpc,Mcore/core.proto=github.com/letsencrypt/boulder/rpc/pb/core:. va/va.proto
